Implements #79, #83
git-svn-id: http://www.phpbms.org/svn/trunk@200 66a7d52b-23fa-0310-b706-d52274950f5d

+++ phpbms/search.php
@@ -124,7 +124,10 @@
 		case "delete":
 			//=====================================================================================================
 			$theids=explode(",",$_POST["theids"]);
-			$tempmessage=delete_record($theids);
+			if(function_exists("delete_record"))
+				$tempmessage=delete_record($theids);
+			else
+				$tempmessage="Delete function not defined.";
 			if($tempmessage) $statusmessage=$tempmessage;
 		break;
 		case "advanced search":
